fix errors and small refactor
