+++ logging_reader/src/main/AndroidManifest.xml
@@ -4,7 +4,7 @@
     <application>
         <provider
             android:name="androidx.core.content.FileProvider"
-            android:authorities="ru.touchin.roboswag.core.log.fileprovider"
+            android:authorities="${packageName}.fileprovider"
             android:exported="false"
             android:grantUriPermissions="true">
             <meta-data
