+++ logging_reader/src/main/java/ru/touchin/roboswag/loggging_reader/DebugLogsDialogFragment.kt
@@ -1,4 +1,4 @@
-package ru.touchin.roboswag.core.log_file
+package ru.touchin.roboswag.loggging_reader
 
 import android.content.Intent
 import android.os.Bundle
@@ -9,11 +9,9 @@ import android.widget.AdapterView
 import android.widget.ArrayAdapter
 import androidx.core.content.FileProvider
 import androidx.fragment.app.DialogFragment
-import androidx.recyclerview.widget.LinearLayoutManager
-import ru.touchin.roboswag.core.log.BuildConfig
-import ru.touchin.roboswag.core.log.R
-import ru.touchin.roboswag.core.log.databinding.DialogFragmentDebugLogsBinding
-import ru.touchin.roboswag.core.log_file.LogFileManager
+import ru.touchin.roboswag.logging_reader.BuildConfig
+import ru.touchin.roboswag.logging_reader.R
+import ru.touchin.roboswag.logging_reader.databinding.DialogFragmentDebugLogsBinding
 import java.io.File
 
 class DebugLogsDialogFragment : DialogFragment() {
@@ -32,28 +30,11 @@ class DebugLogsDialogFragment : DialogFragment() {
         super.onViewCreated(view, savedInstanceState)
         initSpinner()
 
-        binding.logsRecycler.layoutManager = LinearLayoutManager(requireContext())
         binding.logsRecycler.adapter = LogItemAdapter(requireContext(), logItemsList)
         updateRecycler()
 
-        binding.shareBtn.setOnClickListener {
-            val files = LogFileManager(requireContext()).getLogDirectory().listFiles()
-
-            files?.firstOrNull()?.let { firstFile ->
-                val uri = FileProvider.getUriForFile(
-                    requireContext(),
-                    BuildConfig.LIBRARY_PACKAGE_NAME + LogFileManager.fileProviderName,
-                    firstFile
-                )
-
-                val intent = Intent(Intent.ACTION_SEND)
-                intent.addFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION)
-                intent.setType("*/*")
-                intent.putExtra(Intent.EXTRA_STREAM, uri)
-                intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
-                startActivity(intent)
-            }
-        }
+        binding.updateBtn.setOnClickListener { updateRecycler() }
+        binding.shareBtn.setOnClickListener { onShareButtonClick() }
     }
 
     private fun initSpinner() {
@@ -61,25 +42,47 @@ class DebugLogsDialogFragment : DialogFragment() {
         val adapter = ArrayAdapter(requireContext(), android.R.layout.simple_spinner_item, priorityTitle)
         adapter.setDropDownViewResource(android.R.layout.simple_dropdown_item_1line)
 
-        binding.priorityFilter.adapter = adapter;
+        binding.priorityFilter.adapter = adapter
         binding.priorityFilter.onItemSelectedListener = object : AdapterView.OnItemSelectedListener {
             override fun onItemSelected(parent: AdapterView<*>, view: View, position: Int, id: Long) {
                 val priority = LogFileManager.Priority.values()[position]
                 LogFileManager(requireContext()).saveLogcatToFile(priority.tag)
+                binding.logsRecycler.postDelayed({ updateRecycler() }, 500)
             }
 
             override fun onNothingSelected(parent: AdapterView<*>) {}
         }
-        binding.updateBtn.setOnClickListener { updateRecycler() }
     }
 
     private fun updateRecycler() {
+        binding.logsRecycler.adapter?.notifyItemRangeRemoved(0, logItemsList.size)
         logItemsList.clear()
         val files = LogFileManager(requireContext()).getLogDirectory().listFiles()
         files?.firstOrNull()?.let { firstFile ->
-            File(firstFile.getAbsolutePath())
-                .useLines { lines -> lines.forEach { logItemsList.add(it) } }
+            File(firstFile.absolutePath)
+                    .useLines { lines -> lines.forEach { logItemsList.add(it) } }
         }
-        binding.logsRecycler.adapter?.notifyDataSetChanged()
+        binding.logsRecycler.adapter?.notifyItemRangeInserted(0, logItemsList.size)
+    }
+
+    private fun onShareButtonClick() {
+        val files = LogFileManager(requireContext()).getLogDirectory().listFiles()
+
+        files?.firstOrNull()?.let { firstFile ->
+            val uri = FileProvider.getUriForFile(
+                    requireContext(),
+                    BuildConfig.LIBRARY_PACKAGE_NAME + LogFileManager.fileProviderName,
+                    firstFile
+            )
+
+            val intent = Intent(Intent.ACTION_SEND).apply {
+                addFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION)
+                type = "*/*"
+                putExtra(Intent.EXTRA_STREAM, uri)
+                flags = Intent.FLAG_ACTIVITY_NEW_TASK
+            }
+            startActivity(intent)
+        }
+
     }
 }
