+++ logging_reader/src/main/res/values/strings.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string name="update_button_title" formatted="false">Update</string>
+    <string name="share_button_title" formatted="false">Share</string>
+</resources>
